Provide a default implementation for ParsedDocstring.to_stan() that it simply calling node2stan.node2stan() and provides caching function.
I needed to move html2stan and flatten function to another, new, module - pydoctor.stanutils - such that we avoid a circular import issue.
